refactor: Moved Logic classes into BudgetBadger.Core project
