+++ BudgetBadger.sln
@@ -28,8 +28,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BudgetBadger.iOS", "src\Bud
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BudgetBadger.Language", "src\BudgetBadger.Language\BudgetBadger.Language.csproj", "{18CA9F3B-33DE-42C0-B701-8B23915E2089}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BudgetBadger.Logic", "src\BudgetBadger.Logic\BudgetBadger.Logic.csproj", "{BCB48C66-ABD9-429B-A53C-B5C68E24BE90}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BudgetBadger.macOS", "src\BudgetBadger.macOS\BudgetBadger.macOS.csproj", "{64193ED8-32BA-4391-8C1A-1C33E31B70AB}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BudgetBadger.Models", "src\BudgetBadger.Models\BudgetBadger.Models.csproj", "{C605AE89-DE39-47B7-974B-69767B9F0690}"
@@ -267,34 +265,6 @@ Global
 		{18CA9F3B-33DE-42C0-B701-8B23915E2089}.Release|x64.Build.0 = Release|Any CPU
 		{18CA9F3B-33DE-42C0-B701-8B23915E2089}.Release|x86.ActiveCfg = Release|Any CPU
 		{18CA9F3B-33DE-42C0-B701-8B23915E2089}.Release|x86.Build.0 = Release|Any CPU
-		{BCB48C66-ABD9-429B-A53C-B5C68E24BE90}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{BCB48C66-ABD9-429B-A53C-B5C68E24BE90}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{BCB48C66-ABD9-429B-A53C-B5C68E24BE90}.Debug|ARM.ActiveCfg = Debug|Any CPU
-		{BCB48C66-ABD9-429B-A53C-B5C68E24BE90}.Debug|ARM.Build.0 = Debug|Any CPU
-		{BCB48C66-ABD9-429B-A53C-B5C68E24BE90}.Debug|ARM64.ActiveCfg = Debug|Any CPU
-		{BCB48C66-ABD9-429B-A53C-B5C68E24BE90}.Debug|ARM64.Build.0 = Debug|Any CPU
-		{BCB48C66-ABD9-429B-A53C-B5C68E24BE90}.Debug|iPhone.ActiveCfg = Debug|Any CPU
-		{BCB48C66-ABD9-429B-A53C-B5C68E24BE90}.Debug|iPhone.Build.0 = Debug|Any CPU
-		{BCB48C66-ABD9-429B-A53C-B5C68E24BE90}.Debug|iPhoneSimulator.ActiveCfg = Debug|Any CPU
-		{BCB48C66-ABD9-429B-A53C-B5C68E24BE90}.Debug|iPhoneSimulator.Build.0 = Debug|Any CPU
-		{BCB48C66-ABD9-429B-A53C-B5C68E24BE90}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{BCB48C66-ABD9-429B-A53C-B5C68E24BE90}.Debug|x64.Build.0 = Debug|Any CPU
-		{BCB48C66-ABD9-429B-A53C-B5C68E24BE90}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{BCB48C66-ABD9-429B-A53C-B5C68E24BE90}.Debug|x86.Build.0 = Debug|Any CPU
-		{BCB48C66-ABD9-429B-A53C-B5C68E24BE90}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{BCB48C66-ABD9-429B-A53C-B5C68E24BE90}.Release|Any CPU.Build.0 = Release|Any CPU
-		{BCB48C66-ABD9-429B-A53C-B5C68E24BE90}.Release|ARM.ActiveCfg = Release|Any CPU
-		{BCB48C66-ABD9-429B-A53C-B5C68E24BE90}.Release|ARM.Build.0 = Release|Any CPU
-		{BCB48C66-ABD9-429B-A53C-B5C68E24BE90}.Release|ARM64.ActiveCfg = Release|Any CPU
-		{BCB48C66-ABD9-429B-A53C-B5C68E24BE90}.Release|ARM64.Build.0 = Release|Any CPU
-		{BCB48C66-ABD9-429B-A53C-B5C68E24BE90}.Release|iPhone.ActiveCfg = Release|Any CPU
-		{BCB48C66-ABD9-429B-A53C-B5C68E24BE90}.Release|iPhone.Build.0 = Release|Any CPU
-		{BCB48C66-ABD9-429B-A53C-B5C68E24BE90}.Release|iPhoneSimulator.ActiveCfg = Release|Any CPU
-		{BCB48C66-ABD9-429B-A53C-B5C68E24BE90}.Release|iPhoneSimulator.Build.0 = Release|Any CPU
-		{BCB48C66-ABD9-429B-A53C-B5C68E24BE90}.Release|x64.ActiveCfg = Release|Any CPU
-		{BCB48C66-ABD9-429B-A53C-B5C68E24BE90}.Release|x64.Build.0 = Release|Any CPU
-		{BCB48C66-ABD9-429B-A53C-B5C68E24BE90}.Release|x86.ActiveCfg = Release|Any CPU
-		{BCB48C66-ABD9-429B-A53C-B5C68E24BE90}.Release|x86.Build.0 = Release|Any CPU
 		{64193ED8-32BA-4391-8C1A-1C33E31B70AB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{64193ED8-32BA-4391-8C1A-1C33E31B70AB}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{64193ED8-32BA-4391-8C1A-1C33E31B70AB}.Debug|ARM.ActiveCfg = Debug|Any CPU
@@ -456,7 +426,6 @@ Global
 		{87CD48D8-EAA2-4112-95F1-02FA96A318E5} = {C43221E6-4877-4ABE-A859-BFAA92D535A3}
 		{FF648A12-E24C-49BD-AC35-8D1ABDC327C4} = {C43221E6-4877-4ABE-A859-BFAA92D535A3}
 		{18CA9F3B-33DE-42C0-B701-8B23915E2089} = {C43221E6-4877-4ABE-A859-BFAA92D535A3}
-		{BCB48C66-ABD9-429B-A53C-B5C68E24BE90} = {C43221E6-4877-4ABE-A859-BFAA92D535A3}
 		{64193ED8-32BA-4391-8C1A-1C33E31B70AB} = {C43221E6-4877-4ABE-A859-BFAA92D535A3}
 		{C605AE89-DE39-47B7-974B-69767B9F0690} = {C43221E6-4877-4ABE-A859-BFAA92D535A3}
 		{6F6B402D-616F-4247-B4C2-B43DB1964B4A} = {F6529E8D-407D-437A-8521-DEE78B7DE45F}
