+++ src/BudgetBadger.Core/Logic/ReportLogic.cs
@@ -2,13 +2,10 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using BudgetBadger.Core.DataAccess;
 using BudgetBadger.Core.LocalizedResources;
-using BudgetBadger.Core.Logic;
 using BudgetBadger.Models;
-using BudgetBadger.Models.Extensions;
 
-namespace BudgetBadger.Logic
+namespace BudgetBadger.Core.Logic
 {
     public class ReportLogic : IReportLogic
     {
