+++ src/BudgetBadger.Core/Logic/TransactionLogic.cs
@@ -4,11 +4,11 @@ using System.Linq;
 using System.Threading.Tasks;
 using BudgetBadger.Core.DataAccess;
 using BudgetBadger.Core.LocalizedResources;
-using BudgetBadger.Core.Logic;
+using BudgetBadger.Core.Utilities;
 using BudgetBadger.Models;
 using BudgetBadger.Models.Extensions;
 
-namespace BudgetBadger.Logic
+namespace BudgetBadger.Core.Logic
 {
     public class TransactionLogic : ITransactionLogic
     {
