+++ src/BudgetBadger.Core/Utilities/CompressionExtensions.cs
