--- src/BudgetBadger.Logic/BudgetBadger.Logic.csproj
@@ -1,17 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
-  </PropertyGroup>
-
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' " />
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' " />
-  <ItemGroup>
-    <ProjectReference Include="..\BudgetBadger.Core\BudgetBadger.Core.csproj" />
-    <ProjectReference Include="..\BudgetBadger.Models\BudgetBadger.Models.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <None Remove="NETStandard.Library" />
-  </ItemGroup>
-</Project>
